src/llm/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
faiss-cpu
openai
python-dotenv
langchain-groq
langchain-huggingface
langchain-classic
langchain-core
langchain-text-splitters
langchain-chroma
langchain-openai
